fix: add extension to longest_increasing_subsequence

+++ my-library/longest_increasing_subsequence.cpp
@@ -1,7 +1,7 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-vector<int> longestNonDecreasingSubsequence(const vector<ll>& nums) {
+vector<int> longestNonDecreasingSubsequence(const vector<long long>& nums) {
     if (nums.empty()) return {};
 
     vector<int> tails;
